Refactor ETE tests into their own files
